--- a/src/main/scala/is/hail/asm4s/Code.scala
+++ b/src/main/scala/is/hail/asm4s/Code.scala
@@ -1,23 +1,25 @@
 package is.hail.asm4s
 
+import is.hail.expr.CM
 import java.lang.reflect.{Constructor, Field, Method, Modifier}
 
 import org.objectweb.asm.Opcodes._
 import org.objectweb.asm.Type
 import org.objectweb.asm.tree._
 
+import scala.collection.generic.Growable
 import scala.reflect.ClassTag
 
 object Code {
   def apply[T](insn: => AbstractInsnNode): Code[T] = new Code[T] {
-    def emit(il: InsnList): Unit = {
-      il.add(insn)
+    def emit(il: Growable[AbstractInsnNode]): Unit = {
+      il += (insn)
     }
   }
 
   def apply[S1, S2](c1: Code[S1], c2: Code[S2]): Code[S2] =
     new Code[S2] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         c1.emit(il)
         c2.emit(il)
       }
@@ -25,7 +27,7 @@ object Code {
 
   def apply[S1, S2, S3](c1: Code[S1], c2: Code[S2], c3: Code[S3]): Code[S3] =
     new Code[S3] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         c1.emit(il)
         c2.emit(il)
         c3.emit(il)
@@ -34,7 +36,7 @@ object Code {
 
   def apply[S1, S2, S3, S4](c1: Code[S1], c2: Code[S2], c3: Code[S3], c4: Code[S4]): Code[S4] =
     new Code[S4] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         c1.emit(il)
         c2.emit(il)
         c3.emit(il)
@@ -44,7 +46,7 @@ object Code {
 
   def apply[S1, S2, S3, S4, S5](c1: Code[S1], c2: Code[S2], c3: Code[S3], c4: Code[S4], c5: Code[S5]): Code[S5] =
     new Code[S5] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         c1.emit(il)
         c2.emit(il)
         c3.emit(il)
@@ -55,7 +57,7 @@ object Code {
 
   def apply[S1, S2, S3, S4, S5, S6](c1: Code[S1], c2: Code[S2], c3: Code[S3], c4: Code[S4], c5: Code[S5], c6: Code[S6]): Code[S6] =
     new Code[S6] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         c1.emit(il)
         c2.emit(il)
         c3.emit(il)
@@ -64,50 +66,139 @@ object Code {
         c6.emit(il)
       }
     }
-}
 
-trait Code[T] {
-  self =>
-  def emit(il: InsnList): Unit
-
-  def empty: Code[Unit] = new Code[Unit] {
-    def emit(il: InsnList): Unit = {
-      // nothing
+  def newInstance[T](parameterTypes: Array[Class[_]], args: Array[Code[_]])(implicit tct: ClassTag[T], tti: TypeInfo[T]): Code[T] = {
+    new Code[T] {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
+        il += (new TypeInsnNode(NEW, Type.getInternalName(tct.runtimeClass)))
+        il += (new InsnNode(DUP))
+        Invokeable.lookupConstructor[T](parameterTypes).invoke(null, args).emit(il)
+      }
     }
   }
 
-  def compare(opcode: Int, rhs: Code[T]): CodeConditional =
+  def newInstance[T]()(implicit tct: ClassTag[T], tti: TypeInfo[T]): Code[T] =
+    newInstance[T](Array[Class[_]](), Array[Code[_]]())
+
+  def newInstance[T, A1](a1: Code[A1])(implicit a1ct: ClassTag[A1],
+    tct: ClassTag[T], tti: TypeInfo[T]): Code[T] =
+    newInstance[T](Array[Class[_]](a1ct.runtimeClass), Array[Code[_]](a1))
+
+  def newInstance[T, A1, A2](a1: Code[A1], a2: Code[A2])(implicit a1ct: ClassTag[A1], a2ct: ClassTag[A2],
+    tct: ClassTag[T], tti: TypeInfo[T]): Code[T] =
+    newInstance[T](Array[Class[_]](a1ct.runtimeClass, a2ct.runtimeClass), Array[Code[_]](a1, a2))
+
+  def newArray[T](size: Code[Int])(implicit tti: TypeInfo[T], atti: TypeInfo[Array[T]]): Code[Array[T]] = {
+    new Code[Array[T]] {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
+        size.emit(il)
+        il += (tti.newArray())
+      }
+    }
+  }
+
+  def whileLoop(condition: Code[Boolean], body: Code[_]): Code[Unit] = {
+    new Code[Unit] {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
+        val l1 = new LabelNode
+        val l2 = new LabelNode
+        il += (l1)
+        condition.emit(il)
+        il += (new LdcInsnNode(0))
+        il += (new JumpInsnNode(IF_ICMPEQ, l2))
+        body.emit(il)
+        il += (new JumpInsnNode(GOTO, l1))
+        il += (l2)
+      }
+    }
+  }
+
+  def invokeStatic[T, S](method: String, parameterTypes: Array[Class[_]], args: Array[Code[_]])(implicit tct: ClassTag[T], sct: ClassTag[S]): Code[S] = {
+    val m = Invokeable.lookupMethod[T, S](method, parameterTypes)
+    assert(m.isStatic)
+    m.invoke(null, args)
+  }
+
+  def invokeStatic[T, S](method: String)(implicit tct: ClassTag[T], sct: ClassTag[S]): Code[S] =
+    invokeStatic[T, S](method, Array[Class[_]](), Array[Code[_]]())
+
+  def invokeStatic[T, A1, S](method: String, a1: Code[A1])(implicit tct: ClassTag[T], sct: ClassTag[S], a1ct: ClassTag[A1]): Code[S] =
+    invokeStatic[T, S](method, Array[Class[_]](a1ct.runtimeClass), Array[Code[_]](a1))(tct, sct)
+
+  def invokeStatic[T, A1, A2, S](method: String, a1: Code[A1], a2: Code[A2])(implicit tct: ClassTag[T], sct: ClassTag[S], a1ct: ClassTag[A1], a2ct: ClassTag[A2]): Code[S] =
+    invokeStatic[T, S](method, Array[Class[_]](a1ct.runtimeClass, a2ct.runtimeClass), Array[Code[_]](a1, a2))(tct, sct)
+
+  def invokeStatic[T, A1, A2, A3, S](method: String, a1: Code[A1], a2: Code[A2], a3: Code[A3])(implicit tct: ClassTag[T], sct: ClassTag[S], a1ct: ClassTag[A1], a2ct: ClassTag[A2], a3ct: ClassTag[A3]): Code[S] =
+    invokeStatic[T, S](method, Array[Class[_]](a1ct.runtimeClass, a2ct.runtimeClass, a3ct.runtimeClass), Array[Code[_]](a1, a2, a3))(tct, sct)
+
+  def _null[T >: Null]: Code[T] = Code(new InsnNode(ACONST_NULL))
+
+  // FIXME: code should really carry around the stack so this type can be correct
+  // Currently, this is a huge potential place for errors.
+  def _pop[T]: Code[T] = Code(new InsnNode(POP))
+
+  // FIXME: code should really carry around the stack so this type can be correct
+  // Currently, this is a huge potential place for errors.
+  def _empty[T]: Code[T] = new Code[T] {
+    def emit(il: Growable[AbstractInsnNode]): Unit = {
+    }
+  }
+
+  def _throw[T <: java.lang.Throwable, U](cerr: Code[T]): Code[U] = Code(cerr, new InsnNode(ATHROW))
+
+  def checkcast[T](v: Code[AnyRef])(implicit tct: ClassTag[T]): Code[T] = Code(
+    v,
+    new TypeInsnNode(CHECKCAST, Type.getInternalName(tct.runtimeClass)))
+
+  def boxBoolean(cb: Code[Boolean]): Code[java.lang.Boolean] = Code.newInstance[java.lang.Boolean, Boolean](cb)
+  def boxInt(ci: Code[Int]): Code[java.lang.Integer] = Code.newInstance[java.lang.Integer, Int](ci)
+  def boxLong(cl: Code[Long]): Code[java.lang.Long] = Code.newInstance[java.lang.Long, Long](cl)
+  def boxFloat(cf: Code[Float]): Code[java.lang.Float] = Code.newInstance[java.lang.Float, Float](cf)
+  def boxDouble(cd: Code[Double]): Code[java.lang.Double] = Code.newInstance[java.lang.Double, Double](cd)
+
+  def booleanValue(x: Code[java.lang.Boolean]): Code[Boolean] = Code.checkcast[java.lang.Boolean](x).invoke[Boolean]("booleanValue");
+  def intValue(x: Code[java.lang.Number]): Code[Int] = Code.checkcast[java.lang.Number](x).invoke[Int]("intValue");
+  def longValue(x: Code[java.lang.Number]): Code[Long] = Code.checkcast[java.lang.Number](x).invoke[Long]("longValue");
+  def floatValue(x: Code[java.lang.Number]): Code[Float] = Code.checkcast[java.lang.Number](x).invoke[Float]("floatValue");
+  def doubleValue(x: Code[java.lang.Number]): Code[Double] = Code.checkcast[java.lang.Number](x).invoke[Double]("doubleValue");
+}
+
+trait Code[+T] {
+  self =>
+  def emit(il: Growable[AbstractInsnNode]): Unit
+
+  def compare[U >: T](opcode: Int, rhs: Code[U]): CodeConditional =
     new CodeConditional {
-      def emitConditional(il: InsnList): (LabelNode, LabelNode) = {
+      def emitConditional(il: Growable[AbstractInsnNode]): (LabelNode, LabelNode) = {
         val ltrue = new LabelNode
         val lfalse = new LabelNode
         self.emit(il)
         rhs.emit(il)
-        il.add(new JumpInsnNode(opcode, ltrue))
-        il.add(new JumpInsnNode(GOTO, lfalse))
+        il += (new JumpInsnNode(opcode, ltrue))
+        il += (new JumpInsnNode(GOTO, lfalse))
         (ltrue, lfalse)
       }
     }
 }
 
 trait CodeConditional extends Code[Boolean] { self =>
-  def emit(il: InsnList): Unit = {
+  def emit(il: Growable[AbstractInsnNode]): Unit = {
     val lafter = new LabelNode
     val (ltrue, lfalse) = emitConditional(il)
-    il.add(lfalse)
-    il.add(new LdcInsnNode(0))
-    il.add(new JumpInsnNode(GOTO, lafter))
-    il.add(ltrue)
-    il.add(new LdcInsnNode(1))
-    il.add(lafter)
+    il += (lfalse)
+    il += (new LdcInsnNode(0))
+    il += (new JumpInsnNode(GOTO, lafter))
+    il += (ltrue)
+    il += (new LdcInsnNode(1))
+    il += (lafter)
   }
 
   // returns (ltrue, lfalse)
-  def emitConditional(il: InsnList): (LabelNode, LabelNode)
+  def emitConditional(il: Growable[AbstractInsnNode]): (LabelNode, LabelNode)
 
   def unary_!(): CodeConditional =
     new CodeConditional {
-      def emitConditional(il: InsnList): (LabelNode, LabelNode) = {
+      def emitConditional(il: Growable[AbstractInsnNode]): (LabelNode, LabelNode) = {
         val (ltrue, lfalse) = self.emitConditional(il)
         (lfalse, ltrue)
       }
@@ -121,12 +212,12 @@ class CodeBoolean(val lhs: Code[Boolean]) extends AnyVal {
 
     case _ =>
       new CodeConditional {
-        def emitConditional(il: InsnList): (LabelNode, LabelNode) = {
+        def emitConditional(il: Growable[AbstractInsnNode]): (LabelNode, LabelNode) = {
           val ltrue = new LabelNode
           val lfalse = new LabelNode
           lhs.emit(il)
-          il.add(new JumpInsnNode(IFEQ, lfalse))
-          il.add(new JumpInsnNode(GOTO, ltrue))
+          il += (new JumpInsnNode(IFEQ, lfalse))
+          il += (new JumpInsnNode(GOTO, ltrue))
           (ltrue, lfalse)
         }
       }
@@ -138,19 +229,25 @@ class CodeBoolean(val lhs: Code[Boolean]) extends AnyVal {
   def mux[T](cthen: Code[T], celse: Code[T]): Code[T] = {
     val cond = lhs.toConditional
     new Code[T] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         val lafter = new LabelNode
         val (ltrue, lfalse) = cond.emitConditional(il)
-        il.add(lfalse)
+        il += (lfalse)
         celse.emit(il)
-        il.add(new JumpInsnNode(GOTO, lafter))
-        il.add(ltrue)
+        il += (new JumpInsnNode(GOTO, lafter))
+        il += (ltrue)
         cthen.emit(il)
         // fall through
-        il.add(lafter)
+        il += (lafter)
       }
     }
   }
+
+  def &&(rhs: Code[Boolean]): Code[Boolean] =
+    Code(lhs, rhs, new InsnNode(IAND))
+
+  def ||(rhs: Code[Boolean]): Code[Boolean] =
+    Code(lhs, rhs, new InsnNode(IOR))
 }
 
 class CodeInt(val lhs: Code[Int]) extends AnyVal {
@@ -170,9 +267,61 @@ class CodeInt(val lhs: Code[Int]) extends AnyVal {
 
   def <=(rhs: Code[Int]): Code[Boolean] = lhs.compare(IF_ICMPLE, rhs)
 
+  def >>(rhs: Code[Int]): Code[Int] = Code(lhs, rhs, new InsnNode(ISHR))
+
+  def <<(rhs: Code[Int]): Code[Int] = Code(lhs, rhs, new InsnNode(ISHL))
+
+  def &(rhs: Code[Int]): Code[Int] = Code(lhs, rhs, new InsnNode(IAND))
+
   def ceq(rhs: Code[Int]): Code[Boolean] = lhs.compare(IF_ICMPEQ, rhs)
 
   def cne(rhs: Code[Int]): Code[Boolean] = lhs.compare(IF_ICMPNE, rhs)
+
+  def toI: Code[Int] = lhs
+  def toL: Code[Long] = Code(lhs, new InsnNode(I2L))
+  def toF: Code[Float] = Code(lhs, new InsnNode(I2F))
+  def toD: Code[Double] = Code(lhs, new InsnNode(I2D))
+}
+
+class CodeLong(val lhs: Code[Long]) extends AnyVal {
+  def +(rhs: Code[Long]): Code[Long] = Code(lhs, rhs, new InsnNode(LADD))
+
+  def -(rhs: Code[Long]): Code[Long] = Code(lhs, rhs, new InsnNode(LSUB))
+
+  def *(rhs: Code[Long]): Code[Long] = Code(lhs, rhs, new InsnNode(LMUL))
+
+  def /(rhs: Code[Long]): Code[Long] = Code(lhs, rhs, new InsnNode(LDIV))
+
+  def compare(rhs: Code[Long]): Code[Int] = Code(lhs, rhs, new InsnNode(LCMP))
+
+  def <(rhs: Code[Long]): Code[Boolean] = compare(rhs) < 0
+  def >(rhs: Code[Long]): Code[Boolean] = compare(rhs) > 0
+
+  def toI: Code[Int] = Code(lhs, new InsnNode(L2I))
+  def toL: Code[Long] = lhs
+  def toF: Code[Float] = Code(lhs, new InsnNode(L2F))
+  def toD: Code[Double] = Code(lhs, new InsnNode(L2D))
+}
+
+class CodeFloat(val lhs: Code[Float]) extends AnyVal {
+  def +(rhs: Code[Float]): Code[Float] = Code(lhs, rhs, new InsnNode(FADD))
+
+  def -(rhs: Code[Float]): Code[Float] = Code(lhs, rhs, new InsnNode(FSUB))
+
+  def *(rhs: Code[Float]): Code[Float] = Code(lhs, rhs, new InsnNode(FMUL))
+
+  def /(rhs: Code[Float]): Code[Float] = Code(lhs, rhs, new InsnNode(FDIV))
+
+  // FIXME: FCMPG vs FCMPL
+  def compare(rhs: Code[Float]): Code[Int] = Code(lhs, rhs, new InsnNode(FCMPG))
+
+  def <(rhs: Code[Float]): Code[Boolean] = compare(rhs) < 0
+  def >(rhs: Code[Float]): Code[Boolean] = compare(rhs) > 0
+
+  def toI: Code[Int] = Code(lhs, new InsnNode(F2I))
+  def toL: Code[Long] = Code(lhs, new InsnNode(F2L))
+  def toF: Code[Float] = lhs
+  def toD: Code[Double] = Code(lhs, new InsnNode(F2D))
 }
 
 class CodeDouble(val lhs: Code[Double]) extends AnyVal {
@@ -198,6 +347,11 @@ class CodeDouble(val lhs: Code[Double]) extends AnyVal {
   def ceq(rhs: Code[Double]): Code[Boolean] = compare(rhs).ceq(0)
 
   def cne(rhs: Code[Double]): Code[Boolean] = compare(rhs).cne(0)
+
+  def toI: Code[Int] = Code(lhs, new InsnNode(D2I))
+  def toL: Code[Long] = Code(lhs, new InsnNode(D2L))
+  def toF: Code[Float] = Code(lhs, new InsnNode(D2F))
+  def toD: Code[Double] = lhs
 }
 
 class CodeArray[T](val lhs: Code[Array[T]])(implicit tti: TypeInfo[T]) {
@@ -206,6 +360,9 @@ class CodeArray[T](val lhs: Code[Array[T]])(implicit tti: TypeInfo[T]) {
 
   def update(i: Code[Int], x: Code[T]): Code[Unit] =
     Code(lhs, i, x, new InsnNode(tti.astoreOp))
+
+  def length(): Code[Int] =
+    Code(lhs, new InsnNode(ARRAYLENGTH))
 }
 
 object Invokeable {
@@ -238,7 +395,8 @@ object Invokeable {
         parameterTypes.map(_.getName).mkString(", ")
       })")
 
-    assert(m.getReturnType == sct.runtimeClass,
+    // generic type parameters return java.lang.Object instead of the correct class
+    assert(m.getReturnType.isAssignableFrom(sct.runtimeClass),
       s"when invoking ${tct.runtimeClass.getName}.$method(): ${m.getReturnType.getName}: wrong return type ${sct.runtimeClass.getName}")
 
     Invokeable(m)
@@ -262,12 +420,13 @@ class Invokeable[T, S](val name: String,
                        val descriptor: String)(implicit tct: ClassTag[T]) {
   def invoke(lhs: Code[T], args: Array[Code[_]]): Code[S] =
     new Code[S] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         if (!isStatic && lhs != null)
           lhs.emit(il)
         args.foreach(_.emit(il))
-        il.add(new MethodInsnNode(invokeOp,
+        il += (new MethodInsnNode(invokeOp,
           Type.getInternalName(tct.runtimeClass), name, descriptor, isInterface))
+        // il += (new TypeInsnNode(CHECKCAST, Type.getInternalName(sct.runtimeClass)))
       }
     }
 }
@@ -285,18 +444,20 @@ object FieldRef {
 class LocalRef[T](i: Int)(implicit tti: TypeInfo[T]) {
   def load(): Code[T] =
     new Code[T] {
-      def emit(il: InsnList): Unit = {
-        il.add(new IntInsnNode(tti.loadOp, i))
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
+        il += (new IntInsnNode(tti.loadOp, i))
       }
     }
 
-  def store(rhs: Code[T]): Code[T] =
-    new Code[T] {
-      def emit(il: InsnList): Unit = {
+  def store(rhs: Code[T]): Code[Unit] =
+    new Code[Unit] {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         rhs.emit(il)
-        il.add(new IntInsnNode(tti.storeOp, i))
+        il += (new IntInsnNode(tti.storeOp, i))
       }
     }
+
+  def storeInsn: Code[Unit] = Code(new IntInsnNode(tti.storeOp, i))
 }
 
 class FieldRef[T, S](f: Field)(implicit tct: ClassTag[T], sti: TypeInfo[S]) {
@@ -308,27 +469,27 @@ class FieldRef[T, S](f: Field)(implicit tct: ClassTag[T], sti: TypeInfo[S]) {
 
   def get(lhs: Code[T]): Code[S] =
     new Code[S] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         if (!isStatic)
           lhs.emit(il)
-        il.add(new FieldInsnNode(getOp,
+        il += (new FieldInsnNode(getOp,
           Type.getInternalName(tct.runtimeClass), f.getName, sti.name))
       }
     }
 
   def put(lhs: Code[T], rhs: Code[S]): Code[Unit] =
     new Code[Unit] {
-      def emit(il: InsnList): Unit = {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
         if (!isStatic)
           lhs.emit(il)
         rhs.emit(il)
-        il.add(new FieldInsnNode(putOp,
+        il += (new FieldInsnNode(putOp,
           Type.getInternalName(tct.runtimeClass), f.getName, sti.name))
       }
     }
 }
 
-class CodeObject[T <: AnyRef](val lhs: Code[T])(implicit tct: ClassTag[T], tti: TypeInfo[T]) {
+class CodeObject[T >: Null](val lhs: Code[T])(implicit tct: ClassTag[T], tti: TypeInfo[T]) {
   def get[S](field: String)(implicit sct: ClassTag[S], sti: TypeInfo[S]): Code[S] =
     FieldRef[T, S](field).get(lhs)
 
@@ -349,4 +510,52 @@ class CodeObject[T <: AnyRef](val lhs: Code[T])(implicit tct: ClassTag[T], tti: 
   def invoke[A1, A2, S](method: String, a1: Code[A1], a2: Code[A2])(implicit a1ct: ClassTag[A1], a2ct: ClassTag[A2],
                                                                     sct: ClassTag[S]): Code[S] =
     invoke[S](method, Array[Class[_]](a1ct.runtimeClass, a2ct.runtimeClass), Array[Code[_]](a1, a2))
+
+  def invoke[A1, A2, A3, S](method: String, a1: Code[A1], a2: Code[A2], a3: Code[A3])
+    (implicit a1ct: ClassTag[A1], a2ct: ClassTag[A2], a3ct: ClassTag[A2], sct: ClassTag[S]): Code[S] =
+    invoke[S](method, Array[Class[_]](a1ct.runtimeClass, a2ct.runtimeClass, a3ct.runtimeClass), Array[Code[_]](a1, a2, a3))
+
+  def ifNull[T](cnullcase: Code[T], cnonnullcase: Code[T]): Code[T] =
+    new Code[T] {
+      def emit(il: Growable[AbstractInsnNode]): Unit = {
+        val lnull = new LabelNode
+        val lafter = new LabelNode
+        lhs.emit(il)
+        il += (new JumpInsnNode(IFNULL, lnull))
+        cnonnullcase.emit(il)
+        il += (new JumpInsnNode(GOTO, lafter))
+        il += (lnull)
+        cnullcase.emit(il)
+        // fall through
+        il += (lafter)
+      }
+    }
+
+  def mapNull[U >: Null](cnonnullcase: Code[U]): Code[U] =
+    ifNull[U](Code._null[U], cnonnullcase)
+
+  // Ideally this would not use a local variable, but I need a richer way to
+  // talk about the way `Code` modifies the stack.
+  def mapNull[U >: Null](cnonnullcase: Code[T] => Code[U]): CM[Code[U]] = for (
+    (stx, x) <- CM.memoize(lhs)
+  ) yield Code(stx,
+    x.ifNull[U](Code._null[U], cnonnullcase(x)))
+
+  def mapNullM[U >: Null](cnonnullcase: Code[T] => CM[Code[U]]): CM[Code[U]] = for (
+    (stx, x) <- CM.memoize(lhs);
+    result <- cnonnullcase(x)
+  ) yield Code(stx,
+    x.ifNull[U](Code._null[U], result))
+
+  // def mapNull[U >: Null](cnonnullcase: Code[T] => Code[U]): Code[U] =
+  //   new Code[U] {
+  //     def emit(il: Growable[AbstractInsnNode]): Unit = {
+  //       val lnull = new LabelNode
+  //       lhs.emit(il)
+  //       il += (new InsnNode(DUP))
+  //       il += (new JumpInsnNode(IFNULL, lnull))
+  //       cnonnullcase(Code.empty[T]).emit(il)
+  //       il += (lnull)
+  //     }
+  //   }
 }