--- a/src/main/scala/is/hail/expr/HailRep.scala
+++ b/src/main/scala/is/hail/expr/HailRep.scala
@@ -1,5 +1,6 @@
 package is.hail.expr
 
+import scala.collection.mutable
 import is.hail.utils.Interval
 import is.hail.variant.{AltAllele, Genotype, Locus, Variant}