It's still broken on a lot of aggregator expressions and it's a big mess, but people can start to take a look.