@danking 2017-01-24
A Hail pipeline:
```
read profile225-with-SEX_FROM_STAT.vds \
filtergenotypes -c ' g.dp > 400 ||
     (g.isHomRef && (g.ad[0] / g.dp < 0.9 || g.gq < 20)) ||
     (g.isHomVar && (g.ad[1] / g.dp < 0.9 || g.pl[0] < 20)) ||
     (g.isHet && ( (g.ad[0] + g.ad[1]) / g.dp < 0.9 || g.ad[1] / g.dp < 0.20 || g.pl[0] < 20 ))' --keep count -g
```

`master`/`c5bdcfd`:
```
hail: info: timing:
  read: 4.795s
  filtergenotypes: 254.277ms
  count: 1m40.8s
  total: 1m45.8s
```
this branch:
```
hail: info: timing:
  read: 5.822s
  filtergenotypes: 617.229ms
  count: 53.545s
  total: 59.984s
```

Ignoring read time: the compiler takes 53% of the time.

![The Flash](https://media.giphy.com/media/lRnUWhmllPI9a/giphy.gif)