@cseed 2017-01-25
Do you want me to look this over or you do have cleanup to do?  I see large blocks of commented code in the FunctionRegistry.  Also, awesome work!