@danking 2017-01-25
I'll do one clean up pass and then we can go through it.