@danking 2017-01-25
It's still a massive pile of code, but I've cleaned it up a lot from where it was.