@danking 2017-03-03
@cseed if this passes then I think it's ready. Plan from here:

1. collect aggregators and place them into a separate method call of the generated class, communicating aggregator results via private class fields
2. add an IL between AST and java byte code.
3. fuse together `CM` and `Code[T]`
4. eliminate null checks from IL

2 and 3 can be done in parallel