@cseed 2017-03-03
I'm going to ask if you can let @jigold's function registry change go in first.  We'd like to include that in the snapshot for the workshop.  That means you'll need to do another rebase.  So sorry.