@cseed 2017-03-03
Thanks @danking!  @jigold's PR is in, so feel free to rebase/merge at your leisure.