@danking 2017-03-03
Test script:

```
from hail import *
hc = HailContext()
(hc.read('/Users/dking/projects/hail-data/profile225-with-SEX_FROM_STAT.vds')
  .filter_genotypes(' g.dp > 400 || \
     (g.isHomRef && (g.ad[0] / g.dp < 0.9 || g.gq < 20)) || \
     (g.isHomVar && (g.ad[1] / g.dp < 0.9 || g.pl[0] < 20)) || \
     (g.isHet && ( (g.ad[0] + g.ad[1]) / g.dp < 0.9 || g.ad[1] / g.dp < 0.20 || \
     g.pl[0] < 20 ))')
  .count(genotypes=True))
```

Mean of thee runs:

 - `master` / `593a469`: 96.66666s
 - this branch: 74.33333s

Which is a 13% reduction in time

The gap has closed from both sides. While rebasing I threw away a bunch of FunctionRegistry entires that had been converted to `Code`. I'll add them back in in a separate PR and see how that changes things again.